requirements.txt
streamlit
pandas
numpy
matplotlib
requests
python-dotenv
openpyxl
